requirements.txt
# Dépendances pour le pipeline ML
numpy>=1.26.0
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
PyYAML==6.0.1
gitpython==3.1.31
streamlit==1.22.0
Pillow==9.5.0